Major upgrade
A complete recreation of this repository.

Initially, the only purpose of this repository was to hold a single Dockerfile, which could be used to create instances of the GCCG server.

However, major changes / fixes had to be made in order to compile the client (and server) to modern computers.
Thus, this repository now includes several directory structures, source codes and other executables.