Temporary Removal of several sections of code
I have completely removed the SDL section in compat.cpp. This is temporary, only so I can get things compiling again.

Updated the Makefile so that we link exclusively to SDL2 libraries. Before, we were linking to both SDl2 and SDL1.2 libraries, which would break the code, and cause it to segfault.
It seems to be kind of working now. It does not crash.